+++ src/app/chatlib/conversation-detail/message/bubble-message/bubble-message.component.ts
@@ -183,35 +183,38 @@ export class BubbleMessageComponent implements OnInit, OnChanges {
   }
 
 
-  onClickOptionsMessage(event, message){
-    this.logger.log('[BUBBLE-MESSAGE] - onClickOptionsMessage', message);
-    this.presentPopover(event, message)
+  // onClickOptionsMessage(event, message){
+  //   this.logger.log('[BUBBLE-MESSAGE] - onClickOptionsMessage', message);
+  //   this.presentPopover(event, message)
+  // }
+  onOptionsMessageFN(event){
+    this.onOptionsMessage.emit(event)
   }
 
 
-  async presentPopover(ev: any, message: MessageModel) {
-    const attributes = {
-      message: message,
-      conversationWith: message.recipient
-   }
-    const popover = await this.popoverController.create({
-      component: BubbleInfoPopoverComponent,
-      cssClass: 'info-popover',
-      componentProps: attributes,
-      event: ev,
-      translucent: true,
-      keyboardClose: true,
-      showBackdrop: false
-    });
-    popover.onDidDismiss().then((dataReturned: any) => {
-      this.logger.log('[BUBBLE-MESSAGE] presentPopover dismissed. Returned value::', dataReturned.data)
-      if(dataReturned.data){
-        this.onOptionsMessage.emit({option: dataReturned.data.option, message: message})
-      }
-    })
+  // async presentPopover(ev: any, message: MessageModel) {
+  //   const attributes = {
+  //     message: message,
+  //     conversationWith: message.recipient
+  //  }
+  //   const popover = await this.popoverController.create({
+  //     component: BubbleInfoPopoverComponent,
+  //     cssClass: 'info-popover',
+  //     componentProps: attributes,
+  //     event: ev,
+  //     translucent: true,
+  //     keyboardClose: true,
+  //     showBackdrop: false
+  //   });
+  //   popover.onDidDismiss().then((dataReturned: any) => {
+  //     this.logger.log('[BUBBLE-MESSAGE] presentPopover dismissed. Returned value::', dataReturned.data)
+  //     if(dataReturned.data){
+  //       this.onOptionsMessage.emit({option: dataReturned.data.option, message: message})
+  //     }
+  //   })
 
-    return await popover.present();
-  }
+  //   return await popover.present();
+  // }
   // ========= begin:: event emitter function ============//
 
   // returnOpenAttachment(event: String) {