+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # chat21-ionic ver 3.0
 
+### 3.0.90-rc.3
+- added: options menu and popover on click to show copy, canned and Json response options
+- changed: icon-button moved outside bubble-message only for the first message of each sender
+
 ### 3.0.90-rc.2
 - added: info and copy icon on received messages
 - added: flash icon to distinguish chat offline message sent also by email